Trait pallet_rewards::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type PalletId: Get<PalletId>;
    type CurrencyId: AssetId + MaxEncodedLen + MaybeSerializeDeserialize + Default;
    type RewardCurrency: Get<Self::CurrencyId>;
    type GroupId: FullCodec + TypeInfo + MaxEncodedLen + Copy + PartialEq + Debug;
    type Currency: MutateHold<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance, Reason = ()> + Mutate<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance> + Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>;
    type RewardMechanism: RewardMechanism;
    type RewardIssuance: RewardIssuance<AccountId = Self::AccountId, CurrencyId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

type PalletId: Get<PalletId>

Identifier of this pallet used as an account where stores the reward that is not claimed. When you distribute reward, the amount distributed goes here.

type CurrencyId: AssetId + MaxEncodedLen + MaybeSerializeDeserialize + Default

Type used to identify currencies.

type RewardCurrency: Get<Self::CurrencyId>

Identifier for the currency used to give the reward.

type GroupId: FullCodec + TypeInfo + MaxEncodedLen + Copy + PartialEq + Debug

Type used to identify groups.

type Currency: MutateHold<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance, Reason = ()> + Mutate<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance> + Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>

Type used to handle currency transfers and reservations.

type RewardMechanism: RewardMechanism

Specify the internal reward mechanism used by this pallet. Check available mechanisms at mechanism module.

type RewardIssuance: RewardIssuance<AccountId = Self::AccountId, CurrencyId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>

Type used to identify the income stream for rewards.