Crate pallet_rewards

Pallet for issuing rewards to block producers.

Re-exports

pub use weights::WeightInfo;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_rewards

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

Pallet rewards for issuing rewards to block producers.

RewardPoint

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

pallet-rewards events

Traits

Config

Configuration trait of this pallet.

OnReward

Hooks to notify when there are any rewards for specific account.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ProposerSubsidyPoints

Block proposer subsidy parameters

RemainingIssuance

Tokens left to issue to farmers at any given time

RewardsEnabled

Whether rewards are enabled

VoterSubsidyPoints

Voter subsidy parameters