Crate pallet_rewards

Pallet for issuing rewards to block producers.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

Pallet rewards for issuing rewards to block producers.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

pallet-rewards events

Traits

Config

Configuration trait of this pallet.

OnReward

Hooks to notify when there are any rewards for specific account.

WeightInfo

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.