Trait pallet_rewards::mechanism::RewardMechanism

pub trait RewardMechanism {
    type Group;
    type Account;
    type Currency;
    type Balance: Balance;
    type DistributionId: DistributionId;
    type MaxCurrencyMovements: Get<u32>;

    fn reward_group(
        group: &mut Self::Group,
        amount: Self::Balance,
        distribution_id: Self::DistributionId
    ) -> Result<(), ArithmeticError>;
    fn deposit_stake(
        account: &mut Self::Account,
        currency: &mut Self::Currency,
        group: &mut Self::Group,
        amount: Self::Balance
    ) -> Result<(), ArithmeticError>;
    fn withdraw_stake(
        account: &mut Self::Account,
        currency: &mut Self::Currency,
        group: &mut Self::Group,
        amount: Self::Balance
    ) -> Result<(), ArithmeticError>;
    fn compute_reward(
        account: &Self::Account,
        currency: &Self::Currency,
        group: &Self::Group
    ) -> Result<Self::Balance, ArithmeticError>;
    fn claim_reward(
        account: &mut Self::Account,
        currency: &Self::Currency,
        group: &Self::Group
    ) -> Result<Self::Balance, ArithmeticError>;
    fn move_currency(
        currency: &mut Self::Currency,
        prev_group: &mut Self::Group,
        next_group: &mut Self::Group
    ) -> Result<(), MoveCurrencyError>;
    fn account_stake(account: &Self::Account) -> Self::Balance;
    fn group_stake(group: &Self::Group) -> Self::Balance;
}

Required Associated Types

type Group

type Account

type Currency

type Balance: Balance

type DistributionId: DistributionId

type MaxCurrencyMovements: Get<u32>

Required Methods

fn reward_group(
group: &mut Self::Group,
amount: Self::Balance,
distribution_id: Self::DistributionId
) -> Result<(), ArithmeticError>

Reward the group mutating the group entity.

fn deposit_stake(
account: &mut Self::Account,
currency: &mut Self::Currency,
group: &mut Self::Group,
amount: Self::Balance
) -> Result<(), ArithmeticError>

Add stake to the account and mutates currency and group to archieve that.

fn withdraw_stake(
account: &mut Self::Account,
currency: &mut Self::Currency,
group: &mut Self::Group,
amount: Self::Balance
) -> Result<(), ArithmeticError>

Remove stake from the account and mutates currency and group to archieve that.

fn compute_reward(
account: &Self::Account,
currency: &Self::Currency,
group: &Self::Group
) -> Result<Self::Balance, ArithmeticError>

Computes the reward for the account

fn claim_reward(
account: &mut Self::Account,
currency: &Self::Currency,
group: &Self::Group
) -> Result<Self::Balance, ArithmeticError>

Claims the reward, mutating the account to reflect this action. Once a reward is claimed, next calls will return 0 until the group will be rewarded again.

fn move_currency(
currency: &mut Self::Currency,
prev_group: &mut Self::Group,
next_group: &mut Self::Group
) -> Result<(), MoveCurrencyError>

Move a currency from one group to another one.

fn account_stake(account: &Self::Account) -> Self::Balance

Returns the balance of an account

fn group_stake(group: &Self::Group) -> Self::Balance

Returns the balance of a group