Trait pallet_rewards::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>;
    type PalletId: Get<PalletId>;
    type DomainId: TypeInfo + MaxEncodedLen + FullCodec + Copy + PartialEq + Debug;
    type CurrencyId: AssetId + MaxEncodedLen;
    type RewardCurrency: Get<Self::CurrencyId>;
    type GroupId: FullCodec + TypeInfo + MaxEncodedLen + Copy + PartialEq + Debug;
    type Currency: MutateHold<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance> + Mutate<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>;
    type RewardMechanism: RewardMechanism;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self, I>> + IsType<<Self as Config>::Event>

type PalletId: Get<PalletId>

Identifier of this pallet used as an account where stores the reward that is not claimed. When you distribute reward, the amount distributed goes here.

type DomainId: TypeInfo + MaxEncodedLen + FullCodec + Copy + PartialEq + Debug

Type used to identify domains.

type CurrencyId: AssetId + MaxEncodedLen

Type used to identify currencies.

type RewardCurrency: Get<Self::CurrencyId>

Identifier for the currency used to give the reward.

type GroupId: FullCodec + TypeInfo + MaxEncodedLen + Copy + PartialEq + Debug

Type used to identify groups.

type Currency: MutateHold<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance> + Mutate<Self::AccountId, AssetId = Self::CurrencyId, Balance = <<Self as Config<I>>::RewardMechanism as RewardMechanism>::Balance>

Type used to handle currency transfers and reservations.

type RewardMechanism: RewardMechanism

Specify the internal reward mechanism used by this pallet. Check available mechanisms at mechanism module.